Adds clearContents for common storage entity behavior compatibility
Adds support for Clearble to sophisticated storage block entities so
that other mods can explicitly clear the contents of an inventory before
destroying the inventory block entity.

Specifically, this will add compatibility with Valkyrien Skies 2 and
stop duplication of all items in sophisticated storage blocks when
assembling or disassembling ships with sophisticated blocks attached to
them.

This change is standalone and requires no support code in
SophisticatedCore