PT_BR Fix
I made a major correction to the localization because the code was not matching the pt_br language. As a result, I had to rewrite the item entries using the format from the previous version to ensure consistency and correctness. Also, when %s%s was being used, the text was displayed incorrectly — in some items, the word was duplicated, while in others it appeared normal.

- Sorry :(